Move `generic_mask` structs and functions into own module
